The gap problem
I've found a pretty good solution. Record last known positions and
last known free gap. When we want to determine if a gap is free,
we calculate the delta of an image within the gap and a frame whose
target position does not intersect with the the end position at the
end of the gap we're interested in. I'll clarify this in docu.